+++ src/Interfaces/user/UserSubscription.ts
@@ -24,7 +24,7 @@ export default interface UserSubscription {
     subscription?: Subscription;
     subscriptionPricingPlan?: SubscriptionPricingPlan;
     subscriptionTiers?: Array<SubscriptionTier>;
-    activeSubscriptionTier?: SubscriptionTier;
+    activeSubscriptionTier?: Array<SubscriptionTier>;
     addons?: Array<UserSubscriptionAddon>;
     highlights?: Array<any>;
     // Accessors